修一修bug

+++ kernel/mem/virtual_memory_manager.cc
@@ -390,7 +390,7 @@ namespace mem
 
         for (a = oldsz; a < newsz; a += PGSIZE)
         {
-            pa = k_pmm.alloc_page();
+            pa = (uint64)k_pmm.alloc_page();
             if (pa == 0)
             {
                 vmfree(pt, oldsz);
